Reorganized structure of the project Addition to HW5
