+++ src/Common/Printing.java
@@ -1,9 +1,9 @@
-package javaBeginners;
+package Common;
 
-public class Print {
+public class Printing {
+
     public static void Print (int t){
         System.out.println("\u001B[36m*** Task number " + t + " ***\u001B[0m");
     }
-
 
 }
