+++ src/Common/Verification.java
@@ -1,4 +1,4 @@
-package javaBeginners;
+package Common;
 
 public class Verification {
