--- src/HW5adv/numberToString.java
@@ -1,29 +1,0 @@
-package HW5adv;
-//https://www.codewars.com/kata/5265326f5fda8eb1160004c8/train/java
-// We need a function that can transform a number into a string.
-
-
-public class numberToString {
-
-
-    public static String numberToString(int num) {
-        // Return a string of the number here!
-       String numberToString = Integer.toString(num);
-
-        return Integer.toString(num);
-    }
-}
-
-/*
-import org.junit.Test;
-import static org.junit.Assert.assertEquals;
-
-public class NumberStringExampleTests {
-  @Test
-  public void tests() {
-    assertEquals("67", Kata.numberToString(67));
-    assertEquals("123", Kata.numberToString(123));
-    assertEquals("999", Kata.numberToString(999));
-  }
-}
- */
