--- src/HW5adv/task1.java
@@ -1,16 +1,0 @@
-package HW5adv;
-
-
-public class task1 {
-    public static void main(String[] args) {
-
-
-        System.out.println(OppositesAttract.isLove(7,8));
-
-        MathOperations.intAddition(5,8);
-
-
-    }
-
-
-}
