--- src/HW5adv/task2.java
@@ -1,6 +1,0 @@
-package HW5adv;
-
-// pyramid
-public class task2 {
-
-}
