--- src/HW5adv/week5.md
@@ -1,17 +1,0 @@
-Week 5
-
-Lessons:
-
-5.0 Unary operators, methods (Sergei)
-https://www.youtube.com/watch?v=gN9mklkJNqA&list=PL6Gtav8N4O7iqcg2kmMCoOv8Xh1PEVYKL&index=18
-
-5.1 Classes and methods (Edward)
-https://www.youtube.com/watch?v=GZg-rZ7kHsc&list=PL6Gtav8N4O7iqcg2kmMCoOv8Xh1PEVYKL&index=19
-
-HW5adv tasks
-https://docs.google.com/document/d/1nBAZxnw0oJO0muZFRsXNsMnAgOFTdNVPJBIpIGWJp1Q/edit?usp=sharing
-https://docs.google.com/document/d/1kfxu0qeNHa6mpAIfcZoZg8azqO140iZmubU3gpWa1lM/edit?usp=sharing
-
-HW5adv video
-https://www.youtube.com/watch?v=tz-ptzDhaTk&list=PL6Gtav8N4O7iqcg2kmMCoOv8Xh1PEVYKL&index=20
-
