+++ src/javaAdvance/HWA5_task2.java
@@ -1,0 +1,6 @@
+package javaAdvance;
+
+// pyramid
+public class HWA5_task2 {
+
+}
