+++ src/javaAdvance/HWA5_tests.java
@@ -1,0 +1,83 @@
+package javaAdvance;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+
+
+
+public class HWA5_tests {
+
+
+}
+
+
+/*
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+import org.junit.runners.JUnit4;
+
+public class SolutionTest {
+    @Test
+    public void testSomething() {
+        assertEquals("", Kata.countingSheep(0));
+        assertEquals("1 sheep...", Kata.countingSheep(1));
+        assertEquals("1 sheep...2 sheep...", Kata.countingSheep(2));
+        assertEquals("1 sheep...2 sheep...3 sheep...", Kata.countingSheep(3));
+    }
+}
+ */
+
+/*
+import org.junit.*;
+import org.junit.rules.*;
+import org.junit.runner.Description;
+
+public class JavaTest{
+  @Test
+  public final void test_all() {
+    Assert.assertEquals( "Nope!" , 4, Java.doubleInteger(2)); // Test cases here!
+  }
+}
+ */
+
+/*
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+
+public class NumberStringExampleTests {
+  @Test
+  public void tests() {
+    assertEquals("67", Kata.numberToString(67));
+    assertEquals("123", Kata.numberToString(123));
+    assertEquals("999", Kata.numberToString(999));
+  }
+}
+ */
+
+/*
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+
+public class OppositesAttractTest {
+
+  @Test
+  public void testOddAndEven() {
+    assertEquals(true, OppositesAttract.isLove(1, 4));
+  }
+
+  @Test
+  public void testEvenAndEven() {
+    assertEquals(false, OppositesAttract.isLove(2, 2));
+  }
+
+  @Test
+  public void testOddAndOdd() {
+    assertEquals(false, OppositesAttract.isLove(1, 1));
+  }
+
+  @Test
+  public void testEvenAndOdd() {
+    assertEquals(true, OppositesAttract.isLove(0, 1));
+  }
+
+}
+ */
