+++ src/javaAdvance/MathOperations.java
@@ -1,0 +1,27 @@
+package javaAdvance;
+
+
+public class MathOperations {
+
+
+    public static double Addition(double a, double b){
+        return a + b;
+    }
+
+    public static double Subtraction(double a, double b){
+        return a - b;
+    }
+
+    public static double Multiplication(double a, double b){
+        return a * b;
+    }
+
+
+    public static double Division(double a, double b){
+        return a / b;
+    }
+
+
+}
+
+
