+++ src/javaAdvance/week5.md
@@ -1,0 +1,43 @@
+Week 5
+
+Lessons:
+
+5.0 Unary operators, methods (Sergei)
+https://www.youtube.com/watch?v=gN9mklkJNqA&list=PL6Gtav8N4O7iqcg2kmMCoOv8Xh1PEVYKL&index=18
+
+5.1 Classes and methods (Edward)
+https://www.youtube.com/watch?v=GZg-rZ7kHsc&list=PL6Gtav8N4O7iqcg2kmMCoOv8Xh1PEVYKL&index=19
+
+javaAdvance tasks
+https://docs.google.com/document/d/1nBAZxnw0oJO0muZFRsXNsMnAgOFTdNVPJBIpIGWJp1Q/edit?usp=sharing
+https://docs.google.com/document/d/1kfxu0qeNHa6mpAIfcZoZg8azqO140iZmubU3gpWa1lM/edit?usp=sharing
+
+javaAdvance video
+https://www.youtube.com/watch?v=tz-ptzDhaTk&list=PL6Gtav8N4O7iqcg2kmMCoOv8Xh1PEVYKL&index=20
+
+HW5:
+
+1. Необходимо написать 4 метода:
+   сложение 2х чисел
+   вычитание 2х чисел
+   умножение 2х чисел
+   деление 2х чисел
+
+
+
+2. https://www.codewars.com/kata/5b077ebdaf15be5c7f000077/train/java
+   Task:
+   Given a non-negative integer, 3 for example, return a string with a murmur:
+    "1 sheep...2 sheep...3 sheep...". Input will always be valid, i.e. no negative integers.
+   
+3. https://www.codewars.com/kata/53ee5429ba190077850011d4/train/java
+   You need to double the integer and return it
+   
+4. https://www.codewars.com/kata/5265326f5fda8eb1160004c8/train/java
+   We need a function that can transform a number into a string.
+
+5. https://www.codewars.com/kata/555086d53eac039a2a000083/train/java
+   Timmy & Sarah think they are in love, but around where they live,they will only know once they
+   pick a flower each. If one of the flowers has an even number of petals and the other has
+   an odd number of petals it means they are in love.Write a function that will take the number
+   of petals of each flower and return true if they are in love and false if they aren't.
